chore: refactoring and split in different files
