+++ 3rdParty/rover_code/rover_simplex.py
@@ -11,7 +11,7 @@ FIELD_WIDTH = 1900
 FIELD_HEIGHT = 1000
 FIELD_BOUNDARY = 200
 DEFAULT_THRESHOLD = 50
-FINISH_VALUE = 239  
+FINISH_VALUE = 239
 MAXIMUM_LOOP_COUNT = 5
 SERVO_UP = 145
 SERVO_DOWN = 0
